==> Gender Biasness/01_preprocessing.ipynb <==
## import packages

import re
import nltk
import numpy as np
import pandas as pd
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer

nltk.download('stopwords')

## load data

df1 = pd.read_csv(r'data\articles1.csv')
df2 = pd.read_csv(r'data\articles2.csv')
df3 = pd.read_csv(r'data\articles3.csv')

df = pd.concat([df1, df2, df3])
df = df.reset_index(drop=True)
del df1, df2, df3

print(df.shape)
df.head()

## data columns

df.columns

## only keeping necessary columns

df = df[['title', 'content', 'author', 'publication']]
print(df.shape)
df.head()

## missing values

df.isnull().sum()

## filling missing values

df['title'] = df['title'].fillna('No Title')
df['author'] = df['author'].fillna('Unknown')

## maximum length of title and content

max_title_length = max(df['title'].apply(lambda x: len(x)))
max_content_length = max(df['content'].apply(lambda x: len(x)))

print('Maximum title length: ', max_title_length)
print('Maximum content length: ', max_content_length)

## sample title and content

print(f'Title: {df["title"][0]}')
print("-------------------------------------------")
print(f'Content: {df["content"][0]}')

# stopwords after removing gender pronouns

stopwords = stopwords.words('english')
gender_pronouns = ['he', 'she', 'her', 'his', 'him', 'hers', 'himself', 'herself', "she's", "he's", "she'd", "he'd", "she'll", "he'll"]
stopwords_ = set(stopwords) - set(gender_pronouns)

## function definition for text preprocessing title and content

def text_preprocessing(text):
    text = text.lower()  # lowercase text
    text = re.sub('[^a-zA-Z]', ' ', text)  # remove non-alphabets
    text = text.split()  # split text
    text = [word for word in text if not word in stopwords_]  # remove stopwords
    text = ' '.join(text)  # join all
    return text

## text preprocessing title and content

df['title'] = df['title'].apply(text_preprocessing)
print('Title text preprocessing completed.')
df['content'] = df['content'].apply(text_preprocessing)
print('Content text preprocessing completed.')

## missing values after text preprocessing

df.isnull().sum()

## lemmatize the text

lemmatizer = WordNetLemmatizer()

def lemmatize_text(text):
    return ' '.join([lemmatizer.lemmatize(w) for w in text.split()])

df['content_lemmatized'] = ''

for i in range(len(df)):
    df['content_lemmatized'][i] = lemmatize_text(df['content'][i])
    if i % 10000 == 0:
        print(i, end=' ')

df.head()

## saving preprocessed data to csv

df.to_csv(r'data\articles_preprocessed.csv', index=False)
