==> Gender Biasness/02_lexicon_sentiment.ipynb <==
## Importing relevant libraries

import pandas as pd
import numpy as np

import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
nltk.download('vader_lexicon')
nltk.download('punkt')

## Stop warnings
import warnings
warnings.filterwarnings('ignore')

## Loading the data

df = pd.read_csv(r'data\articles_preprocessed.csv')
print(df.shape)
df.head()

## Checking for missing values

df.isnull().sum()

## Dropping missing values

df.dropna(inplace=True)
df.reset_index(drop=True, inplace=True)
print(df.shape)

## Function to perform sentiment analysis

def analyze_sentiment(article):
    
    analyzer = SentimentIntensityAnalyzer()
    sentiment_score = analyzer.polarity_scores(article)

    compound_score = sentiment_score['compound']

    if compound_score >= 0.05:
        sentiment = 'positive'
    elif compound_score <= -0.05:
        sentiment = 'negative'
    else:
        sentiment = 'neutral'

    return sentiment, compound_score

## Performing sentiment analysis

df['sentiment'] = np.nan
df['compound_score'] = np.nan

for i in range(len(df)):
    df['sentiment'][i], df['compound_score'][i] = analyze_sentiment(df['content_lemmatized'][i])

    if i % 10000 == 0:
        print(i, sep=' ')

df.head()

## class distribution in the dataset

df['sentiment'].value_counts()

## saving the data

df.to_csv(r'data\articles_sentiment.csv', index=False)
