==> Gender Biasness/03_gender_analysis.ipynb <==
## Importing relevant libraries

import pandas as pd
import numpy as np

## Stop warnings
import warnings
warnings.filterwarnings('ignore')

## Loading the data

df = pd.read_csv(r'data\articles_sentiment.csv')
print(df.shape)
df.head()

## Checking for gendered words in the content column

content = df['content']

print('Occurences of words:')
print('he: ', content.str.contains(' he ').sum())
print('she: ', content.str.contains(' she ').sum())
print('his: ', content.str.contains(' his ').sum())
print('her: ', content.str.contains(' her ').sum())
print('mr: ', content.str.contains(' mr ' or ' mr. ').sum())
print('mrs: ', content.str.contains(' mrs ' or ' mrs. ').sum())

## Listing gendered words

gender_keywords = {'male': [' he ', ' his ', ' him ', ' himself ', ' hes ', ' man ', ' mr. ', ' mr ', ' men '],
                   'female': [' she ', ' her ', ' hers ', ' herself ', ' shes ', ' woman ', ' ms. ', ' ms ', ' mrs. ', ' mrs ', ' women ']}

## Function to figure out which gender the news article is about

def gender_analysis(text):
    
    # Counting the number of gendered words for male and female
    male_count_keyword = sum(text.count(word) for word in gender_keywords['male'])
    female_count_keyword = sum(text.count(word) for word in gender_keywords['female'])

    # Figuring out the dominating gender by comparing the counts
    # A gender is only considered dominating when it is occured at least 25% more frequent than the less frequent gender
    if male_count_keyword >= 1.25 * female_count_keyword:
        return 'male'
    elif female_count_keyword >= 1.25 * male_count_keyword:
        return 'female'
    else:
        return 'neutral'
    

## Applying the function to the content column

df['gender'] = np.nan

for i in range(len(df)):
    df['gender'][i] = gender_analysis(df['content'][i])

df.head()

## gender column value counts

df['gender'].value_counts()

## saving the data

df.to_csv(r'data\articles_gender.csv', index=False)
