==> Gender Biasness/05_word_embeddings copy.ipynb <==
## Importing relevant modules

import numpy as np
import pandas as pd

import gensim.models.keyedvectors as word2vec

## Loading the Word2Vec model stored in Drive

model = word2vec.KeyedVectors.load_word2vec_format(r"data/GoogleNews-vectors-negative300.bin", binary=True)
model.vector_size

## Loading the dataset

df = pd.read_csv(r"data/articles_gender.csv")
print(df.shape)
df.head()

## Function to create sentence embedding

def sentence_embedding(sentence):
    words = sentence.split()
    embedding = np.zeros(300)
    count = 0
    for word in words:
        if word in model.index_to_key:
            embedding += model[word]
            count += 1
    if count > 0:
        embedding /= count
    return embedding

## Creating embedding for the headlines

embeddings = []

for i in range(len(df['title'])):
    embeddings.append(sentence_embedding(df['title'][i]))

    if i % 10000 == 0:
        print(i, end=' ')

embeddings = np.array(embeddings)
embeddings.shape

# Saving the embeddings in the specified file for reusibility

np.save(r"data/embeddings.npy", embeddings)
