==> Gender Biasness/07_machine_learning.ipynb <==
## Importing relevant libraries

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier

from sklearn.metrics import accuracy_score, f1_score

## Connecting drive to colab

from google.colab import drive
drive.mount('/content/drive')

## Loading the dataset and embeddings

embeddings = np.load(r"/content/drive/MyDrive/Colab Notebooks/_Haider/data/embeddings.npy")
df = pd.read_csv(r"/content/drive/MyDrive/Colab Notebooks/_Haider/data/articles_gender.csv")
print(f"Shape of embeddings: {embeddings.shape}")
print(f"Shape of df: {df.shape}")

## Declaring tergets

y_sentiment = df['sentiment'].values
y_gender = df['gender'].values

## Splitting the data into training and testing sets

X_train, X_test, y_sentiment_train, y_sentiment_test, y_gender_train, y_gender_test = train_test_split(embeddings, y_sentiment, y_gender,
                                                                                                       test_size=0.2, random_state=42)

print("X_train shape:", X_train.shape)
print("X_test shape:", X_test.shape)
print("y_sentiment_train shape:", y_sentiment_train.shape)
print("y_sentiment_test shape:", y_sentiment_test.shape)
print("y_gender_train shape:", y_gender_train.shape)
print("y_gender_test shape:", y_gender_test.shape)

## Creating data frame to store results

results_df = pd.DataFrame(columns=['Model', 'Accuracy (Sentiment)', 'F1 Score (Sentiment)', 'Accuracy (Gender)', 'F1 Score (Gender)'])

## scaling the features

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

## KNN Classifier

model_sentiment = KNeighborsClassifier(n_neighbors=7)
model_sentiment.fit(X_train, y_sentiment_train)
y_sentiment_pred = model_sentiment.predict(X_test)

accuracy_sentiment = accuracy_score(y_sentiment_test, y_sentiment_pred)
f1_sentiment = f1_score(y_sentiment_test, y_sentiment_pred, average='weighted')

model_gender = KNeighborsClassifier(n_neighbors=7)
model_gender.fit(X_train, y_gender_train)
y_gender_pred = model_gender.predict(X_test)

accuracy_gender = accuracy_score(y_gender_test, y_gender_pred)
f1_gender = f1_score(y_gender_test, y_gender_pred, average='weighted')

## Adding the performance in the results dataframe

results_df.loc[len(results_df)] = ['KNN', accuracy_sentiment, f1_sentiment, accuracy_gender, f1_gender]
results_df

# Sentiment analysis using Gaussian Naive Bayes
model_sentiment = GaussianNB()
model_sentiment.fit(X_train, y_sentiment_train)
y_sentiment_pred = model_sentiment.predict(X_test)

accuracy_sentiment = accuracy_score(y_sentiment_test, y_sentiment_pred)
f1_sentiment = f1_score(y_sentiment_test, y_sentiment_pred, average='weighted')

# Gender classification using Gaussian Naive Bayes
model_gender = GaussianNB()
model_gender.fit(X_train, y_gender_train)
y_gender_pred = model_gender.predict(X_test)

accuracy_gender = accuracy_score(y_gender_test, y_gender_pred)
f1_gender = f1_score(y_gender_test, y_gender_pred, average='weighted')

# Adding the performance in the results dataframe
results_df.loc[len(results_df)] = ['Gaussian Naive Bayes', accuracy_sentiment, f1_sentiment, accuracy_gender, f1_gender]
results_df

# Sentiment analysis using Decision Tree
model_sentiment = DecisionTreeClassifier()
model_sentiment.fit(X_train, y_sentiment_train)
y_sentiment_pred = model_sentiment.predict(X_test)

accuracy_sentiment = accuracy_score(y_sentiment_test, y_sentiment_pred)
f1_sentiment = f1_score(y_sentiment_test, y_sentiment_pred, average='weighted')

# Gender classification using Decision Tree
model_gender = DecisionTreeClassifier()
model_gender.fit(X_train, y_gender_train)
y_gender_pred = model_gender.predict(X_test)

accuracy_gender = accuracy_score(y_gender_test, y_gender_pred)
f1_gender = f1_score(y_gender_test, y_gender_pred, average='weighted')

# Adding the performance in the results dataframe
results_df.loc[len(results_df)] = ['Decision Tree', accuracy_sentiment, f1_sentiment, accuracy_gender, f1_gender]
results_df

# Sentiment analysis using Random Forest
model_sentiment = RandomForestClassifier(n_estimators=100, random_state=42)
model_sentiment.fit(X_train, y_sentiment_train)
y_sentiment_pred = model_sentiment.predict(X_test)

accuracy_sentiment = accuracy_score(y_sentiment_test, y_sentiment_pred)
f1_sentiment = f1_score(y_sentiment_test, y_sentiment_pred, average='weighted')

# Gender classification using Random Forest
model_gender = RandomForestClassifier(n_estimators=100, random_state=42)
model_gender.fit(X_train, y_gender_train)
y_gender_pred = model_gender.predict(X_test)

accuracy_gender = accuracy_score(y_gender_test, y_gender_pred)
f1_gender = f1_score(y_gender_test, y_gender_pred, average='weighted')

# Adding the performance in the results dataframe
results_df.loc[len(results_df)] = ['Random Forest', accuracy_sentiment, f1_sentiment, accuracy_gender, f1_gender]
results_df

# Sentiment analysis using SVM
model_sentiment = SVC(kernel='linear', random_state=42)
model_sentiment.fit(X_train, y_sentiment_train)
y_sentiment_pred = model_sentiment.predict(X_test)

accuracy_sentiment = accuracy_score(y_sentiment_test, y_sentiment_pred)
f1_sentiment = f1_score(y_sentiment_test, y_sentiment_pred, average='weighted')

# Gender classification using SVM
model_gender = SVC(kernel='linear', random_state=42)
model_gender.fit(X_train, y_gender_train)
y_gender_pred = model_gender.predict(X_test)

accuracy_gender = accuracy_score(y_gender_test, y_gender_pred)
f1_gender = f1_score(y_gender_test, y_gender_pred, average='weighted')

# Adding the performance in the results dataframe
results_df.loc[len(results_df)] = ['SVM', accuracy_sentiment, f1_sentiment, accuracy_gender, f1_gender]
results_df

results_df.to_csv('/content/drive/MyDrive/Colab Notebooks/_Haider/data/ml_results.csv', index=False)
